requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
numpy==1.26.2 